+++ pkgs/markdown/lib/markdown.dart
@@ -46,7 +46,15 @@ library;
 
 import 'src/version.dart';
 
-export 'src/ast.dart';
+export 'src/ast.dart'
+    show
+        Element,
+        LinkBuilder,
+        Node,
+        NodeVisitor,
+        Resolver,
+        Text,
+        UnparsedContent;
 export 'src/block_parser.dart';
 export 'src/block_syntaxes/alert_block_syntax.dart';
 export 'src/block_syntaxes/block_syntax.dart';