+++ pkgs/pool/lib/pool.dart
@@ -153,7 +153,7 @@ class Pool {
   Stream<T> forEach<S, T>(
       Iterable<S> elements, FutureOr<T> Function(S source) action,
       {bool Function(S item, Object error, StackTrace stack)? onError}) {
-    onError ??= (item, e, s) => true;
+    final errorHandler = onError ?? (item, e, s) => true;
 
     var cancelPending = false;
 
@@ -163,7 +163,7 @@ class Pool {
     late Iterator<S> iterator;
 
     Future<void> run(int _) async {
-      while (iterator.moveNext()) {
+      while (!cancelPending && iterator.moveNext()) {
         // caching `current` is necessary because there are async breaks
         // in this code and `iterator` is shared across many workers
         final current = iterator.current;
@@ -182,7 +182,7 @@ class Pool {
         try {
           value = await action(current);
         } catch (e, stack) {
-          if (onError!(current, e, stack)) {
+          if (errorHandler(current, e, stack)) {
             controller.addError(e, stack);
           }
           continue;
@@ -197,11 +197,16 @@ class Pool {
       iterator = elements.iterator;
 
       assert(doneFuture == null);
-      var futures = Iterable<Future<void>>.generate(
-          _maxAllocatedResources, (i) => withResource(() => run(i)));
-      doneFuture = Future.wait(futures, eagerError: true)
-          .then<void>((_) {})
-          .catchError(controller.addError);
+      void forwardFirstError(Object error, StackTrace stack) {
+        if (cancelPending) return;
+        cancelPending = true;
+        controller.addError(error, stack);
+      }
+
+      doneFuture = Iterable<Future<void>>.generate(
+        _maxAllocatedResources,
+        (i) => withResource(() => run(i)).onError<Object>(forwardFirstError),
+      ).wait;
 
       doneFuture!.whenComplete(controller.close);
     }
@@ -210,8 +215,9 @@ class Pool {
       sync: true,
       onListen: onListen,
       onCancel: () async {
-        assert(!cancelPending);
         cancelPending = true;
+        resumeCompleter?.complete();
+        resumeCompleter = null;
         await doneFuture;
       },
       onPause: () {
@@ -303,7 +309,7 @@ class Pool {
 
     Future.sync(onRelease).then((value) {
       _onReleaseCompleters.removeFirst().complete(PoolResource._(this));
-    }, onError: (Object error, StackTrace stackTrace) {
+    }).onError((Object error, StackTrace stackTrace) {
       _onReleaseCompleters.removeFirst().completeError(error, stackTrace);
       _onResourceReleased();
     });